+++ packages/base/src/features/story/components/SpectaMobileListModeContent.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+/** Placeholder for mobile list-story scroll track (desktop: ListStoryVirtualScrollTrack). */
+export function SpectaMobileListModeContent(): JSX.Element {
+  return (
+    <div
+      id="jgis-story-segment-panel"
+      className="jgis-story-viewer-panel jgis-story-viewer-panel-specta-mod-list jgis-story-mobile-list-mode-stub"
+    >
+      <p>List story mobile view (TODO)</p>
+    </div>
+  );
+}
